Use one common DI
This removes the usage of the `DependencyService` and uses `IServiceProvider`. Since there's even another DI container (`Locator`), I've exposed the `IServiceProvider` to `Global`. Once the move to Blazor happens, we can cut out the dependency to it too so that services, views, view models, etc are agnostic of whatever DI is used ( through constructor paramters which are automatically injected when you request an instance through DI)